requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.0
atomicwrites==1.4.0
attrs==19.3.0
awscli==1.18.64
bcrypt==3.1.7
botocore==1.16.14
cffi==1.14.0
click==7.1.2
colorama==0.4.3
distlib==0.3.0
Django==3.0.7
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Script==2.0.6
Flask-Testing==0.8.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
virtualenv==20.0.20
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
